Release session after implicit read, and implement ping
When doing a remote change of IP-address or name, an implicit
read is first done by the IO-controller to read out info.
A session is created for this.
Close the session immediately after an implicit read,
as there is no info whether the session ever will be used again.

There might also be incoming RPC ping requests when changing
the IP-address. Implement ping reponse, and close the session if
the ping is received on a new session.

Improve logging and documentation.

Also the Linux script for changing the IP address etc should be
able to handle invalid (not set) default gatway.

Closes #54