Rework test setup
The setup functions currently allocate a new instance of p-net for
every testsuite. These instances are left hanging around when the
testsuite finishes. There is also a timer which which will
periodically fire and do work on old instances, including any global
state they may access. This may be the cause of the intermittent
segfaults.

Use a static instance of p-net instead. Drive the timer functionality
manually to better control how it fires. This has the added benefit of
speeding up the unit tests since there is no need to let time pass.

Change-Id: I5c067aadc5a5667c9fad9e6211c2f7214ab4d36c