Use cmake to handle build options
Move user configurable build options to options.h, so they can be
set from the command line when running cmake. This makes it easy to
set them from a build script.

Also allow options to be overriden from the compiler when
building. This makes it easier to set the build options in an IDE or
similar.

Setting the build options to reduce footprint is required to be able
to run the tests on xmc48relax.