Functions triggering informative callbacks should be void
The return values from some informative user callbacks should
not affect the stack.
Is it more obvious if the functions triggering them are void.

We could possibly change the public API of the relevant callbacks
later to make them void, but for now use the same return
value meaning for all public callbacks.
Documentation on the public API will be improved in a separate
pullrequest.