snmp: Increase size of sysLocation
The sysLocation variable is now up to 255 bytes
large (up from 22 bytes) and saved to a new file.
The first 22 bytes is the IM_Tag_Location in
I&M 1, as before.
This change is needed in order to pass the
tests for Conformance Class B.

In fspm.c, the set/get functions for
IM_Tag_location now pads string with spaces.

Logging in pf_snmp.c was refactored and
unit tests were added.