lldp: Only save received IPv4 address, not IPv6.
Previously, if a received LLDP packet contained
multiple Management Address TLVs, only the last
one would be saved for later use. In the case
that the first TLV was the IPv4 address
and a second TLV contained an IPv6 address, the
IPv4 address would be discarded.
As Profinet uses Ipv4, this was bad.

Fixed by only saving the IPv4 address.

Unit tests were also added.