rt-kernel: Use lwip hook function
This removes the xmc4-specific hook function
serving the same purpose, replacing it with a
platform-agnostic hook function.

A drawback with this solution is that the frame
receive handler now checks if frame is a Profinet
frame only after it has checked for IP/ARP
(see ethernet_input() in lwip source tree).
This will increase latency somewhat.

The patch file was updated with changes in lwip
to support this:
- LWIP_HOOK_UNKNOWN_ETH_PROTOCOL was enabled.
- An API function for setting hook function was added.
  If no such hook function is set, received frames
  of unknown type will be dropped.
- To avoid an extra context switch when handling received
  frame, LWIP_TCPIP_CORE_LOCKING_INPUT was activated.