Added destructor for pnet_t members
This is a destructor for pnet_init_only(), meaning it does not free the
pnet_t structure itself. It will only free the dynamically allocated members of pnet_t.

Implementing a pnet_init() destructor in addition to this would be
trivial.

Seems like this has a use in tests' TearDown() function to release
dynamically allocated objects.