DCP identify return false if it hasn't handled a request
The function pf_dcp_identify_req, would return 1 even if the request
wasn't handled by it, this would lead to replies to invalid requests
sent to the device. Fixing this so that it only returns 1 if the
request being made passes the necessary checks.

Fixed the invalid ident_req array in the gtests, filled with proper
station name, and tests now check if the station name, dcp data length
and dcp block length are correct for packets.