Fix failed test when using --gtest_repeat
Using the feature --gtest_repeat would fail for the dcp test as the new
test introduced would change a byte of the array to make sure that the
device ID was not recognized. But when running the test again on repeat,
the array would stay modified, and the test fail because the ID was
incorrect.