Change API usage order in sample app
The API function pnet_output_get_data_and_iops() has a flag p_new_flag
that will be set to true once a new frame has received. It is
of limited use since it does not take the slot or subslot into account,
it just sets the flag to true the first time we read content from an
incoming data buffer.

But also pnet_input_get_iocs() will reset this flag when reading
the buffer.

Document this behavior, and write an recommendation to call
pnet_output_get_data_and_iops() before pnet_input_get_iocs()
if the value of the flag is important for the application.

Change the exection order of these two functions in the sample app.