Add support for AdjustLinkState in PDPortAdjust
This patch adds rudimentary support for enabling/disabling an Ethernet
port from TIA portal, using the Activate/Deactivate adv. port setting.
Rudimentary because it only supports Linux atm.

Note, in pf_pdport_ar_connect_int() we make sure to re-enable the port
after a reset of all port settings.  Because we may not get a re-enable
command over PROFINET.  This probably applies also to P2P boundary, but
that remains to be verified.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>